Decoration position to float

+++ src/Engine/Events/Processor.cpp
@@ -59,14 +59,14 @@ void checkDecorationEvents() {
         const LevelDecoration &decoration = pLevelDecorations[decorationId];
 
         if (decoration.uFlags & LEVEL_DECORATION_TRIGGERED_BY_TOUCH) {
-            if ((decoration.vPosition - pParty->pos.toInt()).length() < decoration.uTriggerRange) {
+            if ((decoration.vPosition - pParty->pos).length() < decoration.uTriggerRange) {
                 eventProcessor(decoration.uEventID, Pid(OBJECT_Decoration, decorationId), 1);
             }
         }
 
         if (decoration.uFlags & LEVEL_DECORATION_TRIGGERED_BY_MONSTER) {
             for (int i = 0; i < pActors.size(); i++) {
-                if ((decoration.vPosition - pActors[i].pos).length() < decoration.uTriggerRange) {
+                if ((decoration.vPosition.toInt() - pActors[i].pos).length() < decoration.uTriggerRange) {
                     eventProcessor(decoration.uEventID, Pid(), 1);
                 }
             }
@@ -74,7 +74,7 @@ void checkDecorationEvents() {
 
         if (decoration.uFlags & LEVEL_DECORATION_TRIGGERED_BY_OBJECT) {
             for (int i = 0; i < pSpriteObjects.size(); i++) {
-                if ((decoration.vPosition - pSpriteObjects[i].vPosition).length() < decoration.uTriggerRange) {
+                if ((decoration.vPosition.toInt() - pSpriteObjects[i].vPosition).length() < decoration.uTriggerRange) {
                     eventProcessor(decoration.uEventID, Pid(), 1);
                 }
             }
