+++ src/Engine/Graphics/Renderer/BaseRenderer.cpp
@@ -243,7 +243,7 @@ void BaseRenderer::PrepareDecorationsRenderList_ODM() {
         }
 
         // view cull
-        if (!IsCylinderInFrustum(pLevelDecorations[i].vPosition.toFloat(), 512.0f)) continue;
+        if (!IsCylinderInFrustum(pLevelDecorations[i].vPosition, 512.0f)) continue;
 
         // LevelDecoration *decor = &pLevelDecorations[i];
         if ((!(pLevelDecorations[i].uFlags & LEVEL_DECORATION_OBELISK_CHEST) ||
@@ -296,7 +296,7 @@ void BaseRenderer::PrepareDecorationsRenderList_ODM() {
                                 color = colorTable.White;
                             }
                         }
-                        pStationaryLightsStack->AddLight(pLevelDecorations[i].vPosition.toFloat() +
+                        pStationaryLightsStack->AddLight(pLevelDecorations[i].vPosition +
                             Vec3f(0, 0, decor_desc->uDecorationHeight / 2),
                             frame->uGlowRadius, color, _4E94D0_light_type);
                     }  // for light
